refactor logger: move some independent methods into separate util file
