+++ src/modules/logger/CMakeLists.txt
@@ -42,6 +42,7 @@ px4_add_module(
 		log_writer.cpp
 		log_writer_file.cpp
 		log_writer_mavlink.cpp
+		util.cpp
 		watchdog.cpp
 	DEPENDS
 		version
